+++ R/geom-label-pairwise.r
@@ -24,6 +24,7 @@ geom_label_pairwise <-
            label.r = grid::unit(0.15, "lines"),
            segment.linewidth = 0.5,
            arrow = NULL,
+           size.unit = "mm",
            na.rm = FALSE,
            show.legend = FALSE,
            inherit.aes = FALSE) {
@@ -65,6 +66,7 @@ geom_label_pairwise <-
       label.r = label.r,
       segment.linewidth = segment.linewidth,
       arrow = arrow,
+      size.unit = size.unit,
       na.rm = na.rm,
       ...
     )
@@ -98,6 +100,7 @@ GeomLabelPairwise <-
                    draw_panel = function(data, panel_params, coord, #panel_scales,
                                          parse = FALSE,
                                          na.rm = FALSE,
+                                         size.unit = "mm",
                                          default.colour = "black",
                                          colour.target = "all",
                                          default.alpha = 1,
@@ -142,6 +145,9 @@ GeomLabelPairwise <-
                                         just = data$hjust,
                                         a = "x", b = "y")
                      }
+
+                     size.unit <- resolve_text_unit(size.unit)
+
                      if (!inherits(label.padding, "margin")) {
                        label.padding <- rep(label.padding, length.out = 4)
                      }
@@ -177,7 +183,7 @@ GeomLabelPairwise <-
                                                 col = ifelse(any(colour.target %in% c("all", "text")),
                                                              ggplot2::alpha(row$colour, text.alpha),
                                                              ggplot2::alpha(default.colour, text.alpha)),
-                                                fontsize = row$size * .pt,
+                                                fontsize = row$size * size.unit,
                                                 fontfamily = row$family,
                                                 fontface = row$fontface,
                                                 lineheight = row$lineheight
@@ -227,5 +233,5 @@ GeomLabelPairwise <-
 
                    },
 
-                   draw_key = draw_key_text
+                   draw_key = ggplot2::draw_key_label
   )